--- threadly-apps/app-api/src/main/java/com/threadly/auth/exception/UserAuthenticationException.java
@@ -1,21 +1,0 @@
-package com.threadly.auth.exception;
-
-import com.threadly.ErrorCode;
-import org.springframework.security.core.AuthenticationException;
-
-/**
- * 사용자 인증 관련 예외
- */
-public class UserAuthenticationException extends AuthenticationException {
-
-  ErrorCode errorCode;
-
-  public UserAuthenticationException(ErrorCode errorCode) {
-    super(errorCode.getDesc());
-    this.errorCode = errorCode;
-  }
-
-  public ErrorCode getErrorCode() {
-    return errorCode;
-  }
-}
