--- threadly-apps/app-api/src/main/java/com/threadly/controller/auth/request/UserLoginRequest.java
@@ -1,21 +1,0 @@
-package com.threadly.controller.auth.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * 로그인 요청 DTO
- */
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserLoginRequest {
-
-  private String email;
-  private String password;
-
-
-}
