--- threadly-apps/app-api/src/main/java/com/threadly/controller/post/request/CreatePostCommentRequest.java
@@ -1,12 +1,0 @@
-package com.threadly.controller.post.request;
-
-/**
- * 게시글 댓글 생성 요청 DTO
- * @param content
- */
-public record CreatePostCommentRequest(
-    String content
-
-) {
-
-}
