--- threadly-apps/app-api/src/main/java/com/threadly/controller/post/request/CreatePostRequest.java
@@ -1,13 +1,0 @@
-package com.threadly.controller.post.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-/**
- * 게시글 작성 요청 DTO
- */
-
-public record CreatePostRequest(@NotBlank @NotNull String content) {
-
-
-}
