--- threadly-apps/app-api/src/main/java/com/threadly/controller/user/UserController.java
@@ -1,87 +1,0 @@
-package com.threadly.controller.user;
-
-import com.threadly.auth.verification.EmailVerificationUseCase;
-import com.threadly.controller.user.mapper.UserProfileRequestMapper;
-import com.threadly.controller.user.request.UserProfileRequest;
-import com.threadly.controller.user.request.UserRegisterRequest;
-import com.threadly.user.RegisterUserUseCase;
-import com.threadly.user.UpdateUserUseCase;
-import com.threadly.user.command.UserRegistrationCommand;
-import com.threadly.user.response.UserProfileApiResponse;
-import com.threadly.user.response.UserRegistrationResponse;
-import jakarta.validation.Valid;
-import java.net.URI;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/users")
-public class UserController {
-
-  private final RegisterUserUseCase registerUserUseCase;
-  private final EmailVerificationUseCase emailVerificationUseCase;
-  private final UpdateUserUseCase updateUserUseCase;
-
-  private final UserProfileRequestMapper userProfileRequestMapper;
-
-
-  /**
-   * 회원 가입
-   *
-   * @param request
-   * @return
-   */
-  @PostMapping("")
-  public UserRegistrationResponse register(
-      @Valid @RequestBody UserRegisterRequest request
-  ) {
-
-    /*회원 가입*/
-    UserRegistrationResponse response = registerUserUseCase.register(
-        UserRegistrationCommand.builder()
-            .email(request.getEmail())
-            .userName(request.getUserName())
-            .password(request.getPassword())
-            .phone(request.getPhone())
-            .build()
-    );
-
-    /*인증 메일 전송*/
-    emailVerificationUseCase.sendVerificationEmail(response.getUserId());
-
-    return response;
-  }
-
-  @PostMapping("/profile")
-  public ResponseEntity<UserProfileApiResponse> setUserProfile(
-      @RequestBody UserProfileRequest request) {
-    /*userId 추출*/
-    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-    String userId = (String) auth.getCredentials();
-
-    URI location = URI.create("/api/users/" + userId);
-
-    return ResponseEntity.created(location)
-        .body(updateUserUseCase.upsertUserProfile(userId, userProfileRequestMapper.toCommand(request)));
-  }
-
-  /**
-   * test - 사용자 정보 비밀번호 수정
-   *
-   * @return
-   */
-  @PostMapping("/update/password")
-  public boolean updatePassword() {
-    return true;
-
-  }
-
-
-}
