+++ threadly-apps/app-api/src/main/java/com/threadly/global/exception/TokenAuthenticationException.java
@@ -1,0 +1,21 @@
+package com.threadly.global.exception;
+
+import com.threadly.ErrorCode;
+import org.springframework.security.core.AuthenticationException;
+
+/**
+ * Token 인증 관련 예외
+ */
+public class TokenAuthenticationException extends AuthenticationException {
+
+  ErrorCode errorCode;
+
+  public TokenAuthenticationException(ErrorCode errorCode) {
+    super(errorCode.getDesc());
+    this.errorCode = errorCode;
+  }
+
+  public ErrorCode getErrorCode() {
+    return errorCode;
+  }
+}
