+++ threadly-apps/app-api/src/main/java/com/threadly/global/filter/FilterBypassMatcher.java
@@ -1,0 +1,26 @@
+package com.threadly.global.filter;
+
+import java.util.List;
+import org.springframework.stereotype.Component;
+import org.springframework.util.AntPathMatcher;
+
+/**
+ * filter 제외 경로
+ */
+public class FilterBypassMatcher {
+
+  public static final List<String> WHITE_LIST = List.of(
+      "/api/users",
+      "/api/auth/verify-email",
+      "/api/auth/login",
+      "/api/auth/reissue"
+  );
+
+  private static final AntPathMatcher PATH_MATCHER = new AntPathMatcher();
+
+  public static boolean shouldBypass(String uri) {
+    return WHITE_LIST.stream().anyMatch(pattern -> PATH_MATCHER.match(pattern, uri));
+  }
+
+
+}
