+++ threadly-apps/app-api/src/main/java/com/threadly/post/request/UpdatePostRequest.java
@@ -1,0 +1,13 @@
+package com.threadly.post.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+/**
+ * 게시글 수정 요청 DTO
+ */
+
+public record UpdatePostRequest(@NotBlank @NotNull String content) {
+
+
+}
