+++ threadly-apps/app-api/src/main/java/com/threadly/user/request/CreateUserProfileRequest.java
@@ -1,0 +1,22 @@
+package com.threadly.user.request;
+
+import com.threadly.user.UserGenderType;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * 사용자 프로필 초기 설정 요청 DTO
+ */
+@Getter
+@Setter
+@AllArgsConstructor
+public class CreateUserProfileRequest {
+
+  private String nickname;
+  private String statusMessage;
+  private String bio;
+  private UserGenderType gender;
+  private String profileImageUrl;
+}
+
