+++ threadly-apps/app-api/src/main/java/com/threadly/user/request/UserRegisterRequest.java
@@ -1,0 +1,29 @@
+package com.threadly.user.request;
+
+import com.threadly.annotation.PasswordEncryption;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class UserRegisterRequest {
+
+  @NotNull
+  @NotBlank
+  private String email;
+
+  @NotNull
+  @NotBlank
+  private String userName;
+
+  @NotNull
+  @NotBlank
+  @PasswordEncryption
+  private String password;
+
+  @NotNull
+  private String phone;
+
+}
