+++ threadly-apps/app-api/src/test/java/com/threadly/BaseApiTest.java
@@ -8,7 +8,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.threadly.controller.auth.request.UserLoginRequest;
+import com.threadly.auth.request.UserLoginRequest;
 import com.threadly.testsupport.fixture.users.UserFixtureLoader;
 import com.threadly.utils.TestLogUtils;
 import java.io.UnsupportedEncodingException;
