+++ threadly-apps/app-api/src/test/java/com/threadly/controller/auth/AuthControllerTest.java
@@ -16,7 +16,7 @@ import com.threadly.CommonResponse;
 import com.threadly.ErrorCode;
 import com.threadly.auth.token.response.LoginTokenResponse;
 import com.threadly.auth.verification.response.PasswordVerificationToken;
-import com.threadly.controller.auth.request.PasswordVerificationRequest;
+import com.threadly.auth.request.PasswordVerificationRequest;
 import com.threadly.testsupport.fixture.users.UserFixtureLoader;
 import com.threadly.utils.TestConstants;
 import java.util.HashMap;
