+++ threadly-apps/app-api/src/test/java/com/threadly/controller/post/BasePostApiTest.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.threadly.BaseApiTest;
 import com.threadly.CommonResponse;
 import com.threadly.auth.token.response.LoginTokenResponse;
-import com.threadly.controller.post.request.CreatePostCommentRequest;
-import com.threadly.controller.post.request.CreatePostRequest;
-import com.threadly.controller.post.request.UpdatePostRequest;
+import com.threadly.post.request.CreatePostCommentRequest;
+import com.threadly.post.request.CreatePostRequest;
+import com.threadly.post.request.UpdatePostRequest;
 import com.threadly.post.comment.create.CreatePostCommentApiResponse;
 import com.threadly.post.comment.get.GetPostCommentListApiResponse;
 import com.threadly.post.create.CreatePostApiResponse;
